Extract kind.go from docker.go.
