+++ pkg/docker/kind.go
@@ -1,0 +1,11 @@
+package docker
+
+func GetKindCidrPrefix() (string, error) {
+	prefix, err := GetDockerCidrPrefix("kind")
+
+	if err == nil {
+		return prefix, nil
+	} else {
+		return "", err
+	}
+}
