+++ pkg/docker/network.go
@@ -67,13 +67,3 @@ func GetDockerCidrPrefix(network string) (string, error) {
 		return "", err
 	}
 }
-
-func GetKindCidrPrefix() (string, error) {
-	prefix, err := GetDockerCidrPrefix("kind")
-
-	if err == nil {
-		return prefix, nil
-	} else {
-		return "", err
-	}
-}
